// relid-device-management/src/uniken/types/rdnaEvents.ts
export interface RDNAError {
  longErrorCode: number;
  shortErrorCode: number;
  errorString: string;
}
// ...
export interface RDNARegisteredDevice {
  devUUID: string;
  devName: string;
  status: string;
  lastAccessedTs: string;
  lastAccessedTsEpoch: number;
  createdTs: string;
  createdTsEpoch: number;
  appUuid: string;
  currentDevice: boolean;
  devBind: number;
}
// ...
export interface RDNAUpdateDeviceDetailsResponseData {
  status_code: number;
  message: string;
  dev_uuid: string;
}

/**
 * RDNA Update Device Details Data
 * Complete response structure for onUpdateDeviceDetails event
 */
export interface RDNAUpdateDeviceDetailsData {
  errCode: number;
  error: RDNAError;
  eMethId: number;
  pArgs: {
    service_details: any;
    response: {
      ResponseData: RDNAUpdateDeviceDetailsResponseData;
      ResponseDataLen: number;
      StatusMsg: string;
      StatusCode: number;
      CredOpMode: number;
    };
    pxyDetails: {
      isStarted: number;
      isLocalhostOnly: number;
      isAutoStarted: number;
      isPrivacyEnabled: number;
      portType: number;
      port: number;
    };
  };
}